Adição da lógica de seleção de receitas

+++ Supervisório Banco Renault/Models/Recipe.cs
@@ -11,19 +11,20 @@ namespace Supervisório_Banco_Renault.Models
 
         public Guid Id { get; set; }
         public string? ModuleCode { get; set; }
-        public int InitialCharacter { get; set; } = 1;
-        public int FinalCharacter { get; set; } = 1;
+        public short InitialCharacter { get; set; } = 1;
+        public short FinalCharacter { get; set; } = 1;
         public bool VerifyModuleTag { get; set; } = true;
         public bool VerifyTraceabilityTag { get; set; } = true;
         public bool VerifyCondenserCovers { get; set; } = true;
         public bool VerifyRadiator { get; set; } = true;
-        public int AteqRadiatorProgram { get; set; } = 1;
-        public float RadiatorPSMinimum { get; set; }
-        public float RadiatorPSMaximum { get; set; }
+        public short AteqRadiatorProgram { get; set; } = 1;
+        public double RadiatorPSMinimum { get; set; }
+        public double RadiatorPSMaximum { get; set; }
         public bool VerifyCondenser { get; set; } = true;
-        public int AteqCondenserProgram { get; set; } = 1;
-        public float CondenserPSMinimum { get; set; }
-        public float CondenserPSMaximum { get; set; }
+        public short AteqCondenserProgram { get; set; } = 1;
+        public double CondenserPSMinimum { get; set; }
+        public double CondenserPSMaximum { get; set; }
+
 
     }
 }
